requirements.txt
numpy==1.26.4
pandas==2.3.0
pillow==11.2.1
streamlit==1.46.0
typing_extensions==4.14.0
streamlit-chatbox==1.1.8
simplejson==3.19.2
opencv-python
openvino

# LangChain Core
langchain
langchain-core
langchain-community

# Google (PaLM) Integration
langchain-google-genai
google-generativeai

# Environment Variable Management
python-dotenv

# Other
tiktoken

# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub

# Pinecone Integration
langchain-pinecone
pinecone